Avoid non-determinstic default keyword argument appearing in documentation.
Whilst working on the Reproducible Builds effort [0], we noticed
that vine could not be built reproducibly. This is due to the use of a
set datatype in the default keyword arguments - this causes a problem
when generating the documentation as the ordering is non-determinstic.

(Mutable datatypes as default keyword arguments are also frowned upon in
Python)

 [0] https://reproducible-builds.org/

Signed-off-by: Chris Lamb <chris@chris-lamb.co.uk>